MT.NonlinearAlg — Method

NonlinearAlg(; alg = LevenbergMarquardt, μ = 1.0)

returns nl_cache that specifies which non linear solver to use for the inverse problem

Keyword Arguments

alg: NonlinearSolve[@ref] algorithm to be used, defaults to LevenbergMarquardt

μ : regularization weight

MT.Occam — Method

Occam(;μgrid= [0.01, 1e6])

MT.OptAlg — Method

OptAlg(; alg = LBFGS, μ = 1.0, kwargs...)

returns nl_cache that specifies which non linear solver to use for the inverse problem

Keyword Arguments

alg: NonlinearSolve[@ref] algorithm to be used, defaults to LBFGS

μ : regularization weight

MT.construct_mixing_models — Method

construct_mixing_models(params, p_names, ϕ, model_list, mixing_type)

returns a mixing_models type containing all the variables for rock physics modeling

Arguments

params : parameter values for rock physics models

p_names : list of symbols, each associating with each value in params

ϕ : vol fraction of different phases

for mixing_type = single_phase, ϕ = [1] (total vol occupied by one phase)

for mixing_type = HS1962_plus and HS1962_minus, ϕ is a vector of one element that can be varied but should be less than 1. This is the vol melt fraction.

model_list : list of model types required to build up the rock physics model and mixing them

mixing_type : determines how to mix model types. Current mixing laws include

single_phase : no mixing, when there's only a single phase

HS1962_plus : mixing two phases to get the Hashim Strikman upper bound

HS1962_minus : mixing two phases to get the Hashim Strikman lower bound

MAL : mixing the two phases using the Modified Archie's Law

MT.d_sigmoid — Method

d_sigmoid(m): gradient for the transformation from optimization domain to model domain. Used for estimating jacobians, but is also useful in analysing sensitivities.

MT.forward! — Method

forward!(r::response, m::model, ω::Vector{T}) where T <: Union{Float32, Float64}:

updates response r type for the given model m at the frequencies ω

MT.forward — Method

forward(m::model, ω::Vector{T}) where T <: Union{Float32, Float64}:

returns a response for the given model m at the frequencies ω

MT.gaussian_kernel — Function

gaussian_kernel(u, σ² = 2)

return gaussian kernel centered at 0, at points given by u with std deviation of σ

MT.get_appres — Method

get_appres(Z, ω): returns the ρₐ for impedance

MT.get_kde — Method

get_kde(data, xgrid; Κ= gaussian_kernel)

returns distribution of data using kernel density estimation

Arguments

data : 1D vector to evaluate distribution for

xgrid : 1D vector to evaluate distribution on

Keyword Arguments

Κ : kernel used to evaluate density, defaults to gaussian_kernel[@ref]

MT.get_kde_image! — Method

get_kde_image!(fig,
    chain::C,
    mDist::mdist;
    hm_kwargs=(;),
    cb_kwargs=(;),
    K=gaussian_kernel,
    half_space_depth=nothing,
    kde_transformation_fn = identity,
    return_kde_mat=false,
    trans_utils=(m=lin_tf, h=lin_tf),
    grid=(m=collect(-1:0.1:5), z=cumsum(mDist.h))) where {
    C <: Chains, mdist <: MTModelDistribution{<:Distribution, <:AbstractArray}}

plots on fig, a heatmap of probability distributions sampled by a chain using kernel density estimation

Arguments

fig : Figure on which the heatmap is plotted

chain : samples in the form Turing.Chains from an MCMC sampling

mDist : apriori model distribution used for MCMC sampling

Keyword Arguments

hm_kwargs : NamedTuple containing keyword arguments for customizing heatmap

cb_kwargs : NamedTuple containing keyword arguments for customizing colorbar

K : kernel used to perform kernel density estimation

half_space_depth : extent of half space, i.e., the last layer, informs how far to extend the half space, defaults to 1.25 × last

kde_transformation_fn : a function that transforms the image domain, eg., use log10 to plot log pdf; defaults to identity which implies no bounds_transformation

return_kde_mat : whether to return the matrix containing the values of heatmap along with corresponding x,y axes; defaults to false

trans_utils : NamedTuple containing functions to transform the samples; defaults to no lin_tf for all parameters

grid : NamedTuple containing grid to evaluate the kernel density on. m refers to the points to evaluate kde of model parameters, z refers to the depth points at which the model samples are inferred, not used if h is not sampled.

Note

Also check relevant tutorial page!

MT.get_kde_image! — Method

get_kde_image!(fig,
    chain,
    mDist;
    hm_kwargs=(;),
    cb_kwargs=(;),
    K=gaussian_kernel,
    half_space_depth=nothing,
    kde_transformation_fn = identity,
    return_kde_mat=false,
    trans_utils=(m=lin_tf, h=lin_tf),
    grid=(m=collect(-1:0.1:5), z=cumsum(mDist.h))) where {
    C <: Chains, mdist <: MTModelDistribution}

plots on fig, a heatmap of probability distributions sampled by a chain using kernel density estimation

Arguments

fig : Figure on which the heatmap is plotted

chain : samples in the form Turing.Chains from an MCMC sampling

mDist : apriori model distribution used for MCMC sampling

Keyword Arguments

hm_kwargs : NamedTuple containing keyword arguments for customizing heatmap

cb_kwargs : NamedTuple containing keyword arguments for customizing colorbar

K : kernel used to perform kernel density estimation

half_space_depth : extent of half space, i.e., the last layer, informs how far to extend the half space, defaults to 1.25 × last

kde_transformation_fn : a function that transforms the image domain, eg., use log10 to plot log pdf; defaults to identity which implies no bounds_transformation

return_kde_mat : whether to return the matrix containing the values of heatmap along with corresponding x,y axes; defaults to false

trans_utils : NamedTuple containing functions to transform the samples; defaults to no lin_tf for all parameters

grid : NamedTuple containing grid to evaluate the kernel density on. m refers to the points to evaluate kde of model parameters, z refers to the depth points at which the model samples are inferred, not used if h is not sampled.

Note

Also check relevant tutorial page!

MT.get_kde_image — Method

get_kde_image!(fig,
    chain,
    mDist;
    hm_kwargs=(;),
    cb_kwargs=(;),
    K=gaussian_kernel,
    half_space_depth=nothing,
    kde_transformation_fn = identity,
    return_kde_mat=false,
    trans_utils=(m=lin_tf, h=lin_tf),
    grid=(m=collect(-1:0.1:5), z=cumsum(mDist.h)))

returns fig, a heatmap of probability distributions sampled by a chain using kernel density estimation

Arguments

fig : Figure on which the heatmap is plotted

chain : samples in the form Turing.Chains from an MCMC sampling

mDist : apriori model distribution used for MCMC sampling

Keyword Arguments

hm_kwargs : NamedTuple containing keyword arguments for customizing heatmap

cb_kwargs : NamedTuple containing keyword arguments for customizing colorbar

K : kernel used to perform kernel density estimation

half_space_depth : extent of half space, i.e., the last layer, informs how far to extend the half space, defaults to 1.25 × last

kde_transformation_fn : a function that transforms the image domain, eg., use log10 to plot log pdf; defaults to identity which implies no bounds_transformation

return_kde_mat : whether to return the matrix containing the values of heatmap along with corresponding x,y axes; defaults to false

trans_utils : NamedTuple containing functions to transform the samples; defaults to no lin_tf for all parameters

grid : NamedTuple containing grid to evaluate the kernel density on. m refers to the points to evaluate kde of model parameters, z refers to the depth points at which the model samples are inferred, not used if h is not sampled.

Note

Also check relevant tutorial page!

MT.get_mean_std_image! — Method

get_mean_std_image!(ax,
    chain,
    mDist::mdist;
    confidence_interval=0.95,
    half_space_depth=nothing,
    plot_kwargs=nothing,
    trans_utils=(m=lin_tf, h=lin_tf))

plots on ax, a bounds plot (using mean and std deviation) of probability distributions sampled by a chain using kernel density estimation

Arguments

fig : Axis on which the probability bounds are plotted

chain : samples in the form Turing.Chains from an MCMC sampling

mDist : apriori model distribution used for MCMC sampling

Keyword Arguments

confidence_interval : a confidence_interval of 0.9 implies 90% of values are within the bounds

half_space_depth : extent of half space, i.e., the last layer, informs how far to extend the half space, defaults to 1.25 × last

plot_kwargs : NamedTuple containing keyword arguments for plots

return_kde_mat : whether to return the matrix containing the values of heatmap along with corresponding x,y axes; defaults to false

trans_utils : NamedTuple containing functions to transform the samples; defaults to no lin_tf for all parameters

z_points : depth points at which bounds are plotted, not used if h is not sampled; defaults to depths corresponding to mean(h)

Note

Also check relevant tutorial page!

MT.get_mean_std_image! — Method

get_mean_std_image!(ax,
    chain,
    mDist;
    confidence_interval=0.95,
    half_space_depth=nothing,
    plot_kwargs=nothing,
    trans_utils=(m=lin_tf, h=lin_tf))

plots on ax, a bounds plot (using mean and std deviation) of probability distributions sampled by a chain using kernel density estimation

Arguments

fig : Axis on which the probability bounds are plotted

chain : samples in the form Turing.Chains from an MCMC sampling

mDist : apriori model distribution used for MCMC sampling

Keyword Arguments

confidence_interval : a confidence_interval of 0.9 implies 90% of values are within the bounds

half_space_depth : extent of half space, i.e., the last layer, informs how far to extend the half space, defaults to 1.25 × last

plot_kwargs : NamedTuple containing keyword arguments for plots

return_kde_mat : whether to return the matrix containing the values of heatmap along with corresponding x,y axes; defaults to false

trans_utils : NamedTuple containing functions to transform the samples; defaults to no lin_tf for all parameters

z_points : depth points at which bounds are plotted, not used if h is not sampled; defaults to depths corresponding to mean(h)

Note

Also check relevant tutorial page!

MT.get_mean_std_image — Method

get_mean_std_image!(chain,
    mDist;
    confidence_interval=0.95,
    half_space_depth=nothing,
    plot_kwargs=nothing,
    trans_utils=(m=lin_tf, h=lin_tf))

return fig, a figure with a bounds plot (using mean and std deviation) of probability distributions sampled by a chain using kernel density estimation

Arguments

fig : Axis on which the probability bounds are plotted

chain : samples in the form Turing.Chains from an MCMC sampling

mDist : apriori model distribution used for MCMC sampling

Keyword Arguments

confidence_interval : a confidence_interval of 0.9 implies 90% of values are within the bounds

half_space_depth : extent of half space, i.e., the last layer, informs how far to extend the half space, defaults to 1.25 × last

plot_kwargs : NamedTuple containing keyword arguments for plots

return_kde_mat : whether to return the matrix containing the values of heatmap along with corresponding x,y axes; defaults to false

trans_utils : NamedTuple containing functions to transform the samples; defaults to no lin_tf for all parameters

z_points : depth points at which bounds are plotted, not used if h is not sampled; defaults to depths corresponding to mean(h)

Note

Also check relevant tutorial page!

MT.get_model_list — Method

get_model_list(chains::chain, mDist::mdist; 
    trans_utils = (m = pow_tf, h = lin_tf,)) where {mdist <: AbstractModelDistribution, chain <: Chains}

returns a list of models from the Chains variable obtained from stochastic_inverse

Arguments

chains : Chains object obtained from the Turing model

mDist : a priori distribution defined before performing stochastic inversion

MT.get_phase — Method

get_phase(Z): returns the phase for impedance

MT.get_ρ_at_z — Method

get_ρ_at_z(pred, zs)

returns the values of the model splatted as a vector in pred at points defined by zs

MT.inverse! — Method

function inverse!(mₖ::model,
        robs::response,
        vars::Vector{Float64},
        alg_cache::occam_cache;
        W= nothing,
        L= nothing,
        max_iters= 30, χ2=1.,
        response_fields::Vector{Symbol}= [k for k ∈ fieldnames(typeof(robs))],
        model_fields::Vector{Symbol}= [k for k ∈ fieldnames(typeof(m₀))], # this will not be used but for the sake of generality for all inverse algs
        model_trans_utils::transform_utils= sigmoid_tf,
        verbose= true
    ):

updates mₖ using occam iteration to fit robs within a misfit of χ2, by default set to 1.0

Variables:

mₖ: Initial model guess, will be updated during the inverse process

robs: response to invert for

vars: variables required for forward modeling, eg., ω for MT

alg_cache: deterimines the algorithm to be performed for inversion

W: Weight matrix, defaults to identity matrix I

L: Regularization matrix, defaults to derivative matrix, given by ∂(@ref)

max_iters= 30: maximum number of iterations

χ2=1.: target misfit

`response_fields: choose data of response to perform inversion on, eg., ρₐ for MT, by default chooses all the data (ρₐ and ϕ)

model_trans_utils:transform_utils= pow_sigmoid_tf: conversion to and from computational domain,

response_trans_utils: for scaling the response parameters,

mᵣ: model in physical domain to be regularized against

reg_term: (For internals) When model in physical domain does not exist, reg_term helps, eg. case of RTO-TKO

verbose: whether to print updates after each iteration, defaults to true

Returns:

return message in the form of return_code and updates mₖ in-place.

Example:

inverse!(m_occam, r_obs, Occam([1e-2, 1e6]))

MT.inverse! — Method

function inverse!(mₖ::model1,
        robs::response,
        vars::Vector{Float64},
        alg_cache::nl_cache;
        W=nothing,
        L=nothing,
        max_iters=30,
        χ2=1.0,
        response_fields::Vector{Symbol}=[k for k in fieldnames(typeof(robs))],
        # model_fields::Vector{Symbol}=[k for k in fieldnames(typeof(mₖ))], # this will not be used but for the sake of generality for all inverse algs
        model_trans_utils::transform_utils=pow_sigmoid_tf,
        response_trans_utils::transform_utils=log_tf,
        verbose::Bool=true,
        mᵣ=nothing) where {
        model1 <: AbstractGeophyModel, response <: AbstractGeophyResponse}

updates mₖ using occam iteration to fit robs within a misfit of χ2, by default set to 1.0.

Variables:

mₖ: Initial model guess, will be updated during the inverse process

robs: response to invert for

vars: variables required for forward modeling, eg., ω for MT

alg_cache::occam_cache: deterimines the algorithm to be performed for inversion

W: Weight matrix, will be I if nothing is provided

L: Regularization matrix, defaults to derivative matrix, given by ∂[@ref]

max_iters: maximum number of iterations, defaults to 30

χ2: threshold misfit, defaults to 1.0

response_fields: choose data of response to perform inversion on, eg., ρₐ for MT, by default chooses all the data (ρₐ and ϕ)

model_fields: will generally be fixed, see docs for details

model_trans_utils: conversion to and from computational domain,

response_trans_utils: for scaling the response parameters,

verbose: whether to print updates after each iteration, defaults to true

mᵣ: model in physical domain to be regularized against

Returns:

return message in the form of return_code and updates mₖ in-place.

Example:

inverse!(m_occam, r_obs, ω, NonlinearAlg(; alg = LevenbergMarquardt, μ = 1.0))

MT.inverse! — Method

function inverse!(mₖ::model1,
        robs::response,
        vars::Vector{Float64},
        alg_cache::opt_cache;
        W=nothing,
        L=nothing,
        max_iters=30,
        χ2=1.0,
        response_fields::Vector{Symbol}=[k for k in fieldnames(typeof(robs))],
        # model_fields::Vector{Symbol}=[k for k in fieldnames(typeof(mₖ))], # this will not be used but for the sake of generality for all inverse algs
        model_trans_utils::transform_utils=pow_sigmoid_tf,
        response_trans_utils::transform_utils=log_tf,
        verbose::Bool=true,
        mᵣ=nothing) where {
        model1 <: AbstractGeophyModel, response <: AbstractGeophyResponse}

updates mₖ using occam iteration to fit robs within a misfit of χ2, by default set to 1.0.

Variables:

mₖ: Initial model guess, will be updated during the inverse process

robs: response to invert for

vars: variables required for forward modeling, eg., ω for MT

alg_cache::occam_cache: deterimines the algorithm to be performed for inversion

W: Weight matrix, will be I if nothing is provided

L: Regularization matrix, defaults to derivative matrix, given by ∂[@ref]

max_iters: maximum number of iterations, defaults to 30

χ2: threshold misfit, defaults to 1.0

response_fields: choose data of response to perform inversion on, eg., ρₐ for MT, by default chooses all the data (ρₐ and ϕ)

model_fields: will generally be fixed, see docs for details

model_trans_utils: conversion to and from computational domain,

response_trans_utils: for scaling the response parameters,

verbose: whether to print updates after each iteration, defaults to true

mᵣ: model in physical domain to be regularized against

Returns:

return message in the form of return_code and updates mₖ in-place.

Example:

inverse!(m_occam, r_obs, ω, NonlinearAlg(; alg = LBFGS, μ = 1.0))

MT.inverse_sigmoid — Method

inverse_sigmoid(): get back to the optimization domain from model domain

MT.linsolve! — Method

linsolve!: Performs inv(B)*y using LinearSolve.jl

MT.mcmc_turing — Method

@model function mcmc_turing(
    m_sample::model,
    vars,
    r_obs::NamedTuple,
    err_resp::MTResponse,
    mDist::mdist,
    rDist::rdist;
    response_fields::Vector{Symbol}= [k for k ∈ fieldnames(typeof(rDist))],
    model_fields::Vector{Symbol}= [k for k ∈ fieldnames(typeof(mDist))],
    trans_utils::NamedTuple = (m = log_tf, h = lin_tf)
    ) where {model <: AbstractModel, mdist <: AbstractModelDistribution, rdist <: AbstractResponseDistribution}

makes a Turing.jl model to perform MCMC sampling

Variables:

vars: variables that need to be passed into the forward function along with model to generate a response

r_obs: named tuple containing the observed data, with the same keys as the fields in the corresponding response

err_resp: response variable that contains the errors

mDist: any subtype of AbstractModelDistribution contains the apriori information

rDist: any subtype of AbstractResponseDistribution contains the likelihood information

Keyword/optional arguments

response_fields: which fields in response to invert for

model_fields: fields in model to draw inference on

trans_utils: to transform the model field variables to and from computational (inference) domain

MT.normal_dist — Method

normal_dist(resp::AbstractVector, err_resp::AbstractVector): returns a multivariate normal distribution centered around resp with std dev err_resp

MT.occam_step! — Method

function occam_step!(mₖ₊₁::model,
    respₖ₊₁::response,
    vars::Union{AbstractVector{Float32}, AbstractVector{Float64}},
    χ2::Union{Float64, Float32},
    μgrid::Vector{Float64},
    lin_utils::linear_utils,
    inv_utils::inverse_utils,
    model_trans_utils::transform_utils,
    response_trans_utils::NamedTuple,
    linsolve_prob::LinearSolve.LinearCache;
    model_fields::Vector{Symbol}= [k for k ∈ fieldnames(typeof(mₖ₊₁))],
    response_fields::Vector{Symbol}= [k for k ∈ fieldnames(typeof(respₖ₊₁))],
    verbose= false
    ):

performs a single step of occam inversion, using golden line search.

Variables:

mₖ: Initial model guess, will be updated during the inverse process

robs: response to invert for

vars: variables required for forward modeling, eg., ω for MT

alg_cache: deterimines the algorithm to be performed for inversion

W: Weight matrix, defaults to identity matrix I

L: Regularization matrix, defaults to discretized derivative matrix given by ∂(@ref)

max_iters: maximum number of iterations, defaults to 30

χ2: target misfit, defaults to 1.0

`response_fields: choose data of response to perform inversion on, eg., ρₐ for MT, by default chooses all the data (ρₐ and ϕ)

model_trans_utils: conversion to and from computational domain,

response_trans_utils: NamedTuple containing transform_utils for scaling different response parameters,

mᵣ: model in physical domain to be regularized against

reg_term: (For internals) When model in physical domain does not exist, reg_term helps, eg. case of RTO-TKO

verbose: whether to print updates after each iteration, defaults to true

MT.stochastic_inverse — Method

stochastic_inverse(
    r_obs::response,
    err_resp::response,
    vars,
    alg_cache::mcmc_cache;
    model_trans_utils::NamedTuple = (m = lin_tf, h = lin_tf)
    )

function to perform sampling

Returns

`AbstractMCMC.jl Chain` containing the vectors used for performing samping. Note that all the variables will be named `var_inf`, for variable inferred. 
The vector in each sample will be all the fields of the model being inferred on concatenated together.

Variables

r_obs: response that needs to inverted for

err_resp: response variable containing the errors associated with observed response

vars: variables that need to be passed into the forward function along with model to generate a response

alg_cache: to tell the compiler what type of stochastic inversion method is to be used

model_trans_utils: A named tuple containing transform_utils for the fields of model that need to be scaled/modified. If not provided for any model field, the field won't be modified.

MT.stochastic_inverse — Method

stochastic_inverse(r_obs::resp1,
        err_resp::resp2,
        vars,
        alg_cache::rto_cache;
        model_trans_utils::NamedTuple=(m=sigmoid_tf, h=lin_tf),
        response_trans_utils::NamedTuple=(ρₐ=lin_tf, ϕ=lin_tf)) where {
        resp1 <: AbstractGeophyResponse, resp2 <: AbstractGeophyResponse}

Returns

`AbstractMCMC.jl Chain` containing the vectors used for performing samping. Note that all the variables will be named `var_inf`, for variable inferred. 
The vector in each sample will be all the fields of the model being inferred on concatenated together.

Arguments

r_obs: response that needs to inverted for

err_resp: response variable containing the errors associated with observed response

vars: variables that need to be passed into the forward function along with model to generate a response

alg_cache: to tell the compiler what type of stochastic inversion method is to be used

Keyword Arguments

model_trans_utils: A named tuple containing transform_utils for the fields of model that need to be scaled/modified. If not provided for any model field, the field won't be modified

response_trans_utils: for scaling the response parameters

MT.uniform_dist — Method

uniform_dist(resp::AbstractVector, err_resp::AbstractVector): returns a uniform normal distribution ∈ [resp-err_resp/2, resp+err_resp/2]

MT.χ² — Method

χ²(dcal::T, dobs::T; W): returns a chi-squared error between the observed and the calculated data. W can optionally be passed to weigh points differently.

MT.∂ — Method

∂(n): returns a nxn matrix for a 1D finite difference stencil using 2 points.

MT.AbstractGeophyModel — Type

Abstract model type that is the supertype of all geophysical models in the package.

MT.AbstractGeophyModelDistribution — Type

Abstract model type that is the supertype of all geophysical model distributions in the package.

MT.AbstractGeophyResponse — Type

Abstract model type that is the supertype of all geophysical responses in the package.

MT.AbstractGeophyResponseDistribution — Type

Abstract model type that is the supertype of all geophysical response distributions in the package.

MT.AbstractModel — Type

Abstract model type that is the supertype of all models in the package.

MT.AbstractModelDistribution — Type

Abstract model type that is the supertype of all model distributions in the package.

MT.AbstractResponse — Type

Abstract model type that is the supertype of all responses in the package.

MT.AbstractResponseDistribution — Type

Abstract model type that is the supertype of all response distributions in the package.

MT.AbstractRockphyModel — Type

Abstract model type that is the supertype of all rock models in the package.

MT.AbstractRockphyModelDistribution — Type

Abstract model type that is the supertype of all rock physics model distributions in the package.

MT.AbstractRockphyResponse — Type

Abstract model type that is the supertype of all rock physics responses in the package.

MT.AbstractRockphyResponseDistribution — Type

Abstract model type that is the supertype of all rock physics response distributions in the package.

MT.Gaillard2008 — Type

Gaillard2008(T)

Electrical conductivity model for melt dependent on Temperature. For further reading look at the references below.

Usage

julia> model = Gaillard2008(1000 + 273.)

julia> log_cond = forward(model, [])

Arguments

T : Temperature of melt (should be greater than 1146.8 K)

References

Gaillard, Fabrice & Malki, Mohammed & Iacono-Marziano, Giada & Pichavant, Michel & Scaillet, Bruno. (2008), "Carbonatite Melts and Electrical Conductivity in the Asthenosphere", Science (New York, N.Y.). 322. 1363-5, doi: 10.1126/science.1164446.

MT.HS1962_minus — Type

HS1962_minus

Hashim-Strikman lower bound for mixing 2 phases

Usage

julia> model_mix = construct_mixing_models([1000. + 273., 2e4],
    [:T, :Ch2o_m],
    [0.1],
    [SEO3, Ni2011],
    [HS1962_minus()]
)
julia> log_cond = forward(model_mix, [])

References

Paul W. J. Glover (2010), "A generalized Archie's law for n phases", Geophysics 2010; 75 (6): E247–E265, doi: https://doi.org/10.1190/1.3509781

MT.HS1962_plus — Type

HS1962_plus

Hashim-Strikman upper bound for mixing 2 phases

Usage

julia> model_mix = construct_mixing_models([1000. + 273., 2e4]
    [:T, :Ch2o_m]],
    Product([Uniform(0., 1.)]),
    [SEO3, Ni2011],
    [HS1962_minus()]
)
julia> log_cond = forward(model_mix, [])

References

Paul W. J. Glover (2010), "A generalized Archie's law for n phases", Geophysics 2010; 75 (6): E247–E265. doi: https://doi.org/10.1190/1.3509781

MT.Jones2012 — Type

Jones2012(T, Ch2o_ol)

Electrical conductivity model for olivine dependent on temperature and water concentration. For further reading look at the references below.

Arguments

T : Temperature of olivine (in K)

Ch2o_ol : water concentration in olivine (in ppm)

Usage

julia> model = Jones2012(1000 + 273., 2e4)

julia> log_cond = forward(model, [])

References

Jones, A. G., J. Fullea, R. L. Evans, and M. R. Muller (2012), "Water in cratonic lithosphere: Calibrating laboratory-determined models of electrical conductivity of mantle minerals using geophysical and petrological observations", Geochem. Geophys. Geosyst., 13, Q06010, doi:10.1029/2012GC004055.

MT.MAL — Type

MAL

Modified Archie's law for mixing 2 phases

Usage

julia> model_mix = construct_mixing_models([1000. + 273., 2e4]
    [:T, :Ch2o_m]],
    Product([Uniform(0., 1.)]),
    [SEO3, Ni2011],
    [MAL(1.2)]
)
julia> log_cond = forward(model_mix, [])

References

Glover, P. W. J., Hole, M. J., & Pous, J. (2000), "A Modified Archie’s Law for two conducting phases", Earth and Planetary Science Letters, 180(3–4), 369–383, doi: https://doi.org/10.1016/S0012-821X(00)00168-0

MT.MTModel — Type

create a model type for a given resistivity distribution that can be used to calculate forward response for 1d MT

MT.MTModelDistribution — Type

mutable struct modelDistribution{T1<: Union{Distribution, AbstractArray}, T2<: Union{Distribution, AbstractArray}} # where T1,T2 
    m::T1
    h::T2
end

create a placeholder to store the Distributions.jl sampler for a priori

MT.MTResponseDistribution — Type

struct responseDistribution{T1<: Union{Function, Nothing}, T2<: Union{Function, Nothing}} # where T1,T2
    ρₐ::T1
    ϕ::T2
end

create a placeholder to store functions to obtain Distributions.jl samplers for the likelihood function

MT.Ni2011 — Type

Ni2011(T, Ch2o_m)

Electrical conductivity model for basaltic melt dependent on Temperature and water content in melt. For further reading look at the references below.

Arguments

T : Temperature of melt (should be greater than 1146.8 K)

Ch2o_m : water concentration in melt (in ppm)

Usage

julia> model = Ni2011(1000 + 273., 2e4)

julia> log_cond = forward(model, [])

References

Ni, H., Keppler, H. & Behrens, H. (2011), "Electrical conductivity of hydrous basaltic melts: implications for partial melting in the upper mantle.", Contrib Mineral Petrol 162, 637–650 (2011), doi: https://doi.org/10.1007/s00410-011-0617-4

MT.Poe2010 — Type

Poe2010(T, Ch2o_ol)

Electrical conductivity model for olivine dependent on temperature and water concentration. For further reading look at the references below.

Arguments

T : Temperature of olivine (in K)

Ch2o_ol : water concentration in olivine (in ppm)

Usage

julia> model = Poe2010(1000 + 273., 2e4)

julia> log_cond = forward(model, [])

References

Brent T. Poe, Claudia Romano, Fabrizio Nestola, Joseph R. Smyth (2010), "Electrical conductivity anisotropy of dry and hydrous olivine at 8GPa", Physics of the Earth and Planetary Interiors,Volume 181, Issues 3–4, 2010, Pages 103-111, ISSN 0031-9201, https://doi.org/10.1016/j.pepi.2010.05.003.

MT.RockphyModelDistribution — Type

RockphyModelDistribution{
    T1 <: Union{Distribution, AbstractArray}, T2 <: Union{Distribution, AbstractArray}} <:
            AbstractRockphyModelDistribution
    params::T1 # vector of parameters 
    p_names::Vector{<:Symbol} # Vector of symbols telling the parameters in vector 
    ϕ::T2 # phase ratios
    model_list::Vector{<:Type}
    mixing_type::Vector

Initializes a placeholder for all the variables to be initialized for use in the stochastic_inverse(@ref stochasticinverse). Has the same structure as `constructmixingmodels`(@ref constructmixing_models) but instead inputs distributions, and therefore respects the same form.

Arguments

params : A distribution from Distributions.jl to sample the rock physics parameters.

p_names : The params vector contains the distributions for the parameters specified by the vector of symbols called p_names

ϕ : Vol. fraction of different phases (can be a distribution or a vector)

model_list : list of models to be used in mixing

mixing_type : Type of mixing model to be used

MT.SEO3 — Type

SEO3(T)

Electrical conductivity model for olivine dependent on temperature.

For further reading look at the references below.

Arguments

- `T` : Temperature of olivine (in K)

Usage

julia> model = SEO3(1000 + 273.)

julia> log_cond = forward(model, [])

References

Constable, S (2006), "SEO3: A new model of olivine electrical conductivity", Geophysical Journal International, Volume 166, Issue 1, July 2006, Pages 435–437, https://doi.org/10.1111/j.1365-246X.2006.03041.x

MT.Sifre2014 — Type

Sifre2014(T, Ch2o_m, Cco2_m)

Electrical conductivity model for melt dependent on Temperature, water content and CO₂ content in melt. For further reading look at the references below.

Arguments

T : Temperature of melt (should be greater than 1146.8 K)

Ch2o_m : water concentration in melt (in ppm)

Cco2_m : Co2 concentration in melt (in ppm)

Usage

julia> model = Sifre2014(1000 + 273., 2e4, 2e4)

julia> log_cond = forward(model, [])

References

Sifré, D., Gardés, E., Massuyeau, M. et al. (2014), "Electrical conductivity during incipient melting in the oceanic low-velocity zone." Nature 509, 81–85 (2014), doi: https://doi.org/10.1038/nature13245

MT.UHO2014 — Type

UHO2014(T, Ch2o_ol)

Electrical conductivity model for olivine dependent on temperature and water concentration. For further reading look at the references below.

Arguments

T : Temperature of olivine (in K)

Ch2o_ol : water concentration in olivine (in ppm)

References

Gardés, E., F. Gaillard, and P. Tarits (2014), "Toward a unified hydrous olivine electrical conductivity law", Geochem. Geophys. Geosyst., 15, 4984–5000, doi:10.1002/2014GC005496.

Usage

julia> model = UHO2014(1000 + 273., 2e4)

julia> log_cond = forward(model, [])

MT.Wang2006 — Type

Wang2006(T, Ch2o_ol)

Electrical conductivity model for olivine dependent on temperature and water concentration. For further reading look at the references below.

Arguments

T : Temperature of olivine (in K)

Ch2o_ol : water concentration in olivine (in ppm)

Usage

julia> model = Wang2006(1000 + 273., 2e4)

julia> log_cond = forward(model, [])

References

Wang, D., Mookherjee, M., Xu, Y. et al. (2006), "The effect of water on the electrical conductivity of olivine", Nature 443, 977–980 (2006), doi: https://doi.org/10.1038/nature05256

MT.Yoshino2009 — Type

Yoshino2009(T, Ch2o_ol)

Electrical conductivity model for olivine dependent on temperature and water concentration. For further reading look at the references below.

Arguments

T : Temperature of olivine (in K)

Ch2o_ol : water concentration in olivine (in ppm)

Usage

julia> model = Yoshino2009(1000 + 273., 2e4)

julia> log_cond = forward(model, [])

References

Takashi Yoshino, Takuya Matsuzaki, Anton Shatskiy, Tomoo Katsura (2009), "The effect of water on the electrical conductivity of olivine aggregates and its implications for the electrical structure of the upper mantle, Earth and Planetary Science Letters", Volume 288, Issues 1–2, 2009, Pages 291-300, ISSN 0012-821X, https://doi.org/10.1016/j.epsl.2009.09.032.

MT.const_matrix — Type

const_matrix(σ)

Fixed electrical conductivity model for the phase.

Arguments

σ : Conductivity of the phase

Usage

julia> model = const_matrix(1000.)

julia> log_cond = forward(model, [])

MT.inverse_utils — Type

struct inverse_utils: contains the utilities for inversion, once initialized, will not be updated in the inversion iterations D: second derivative operator, W: weight matrix, dobs: data response to be inverted for.

MT.linear_utils — Type

struct linear_utils: contains the utilities for linearizing the forward model => mₖ:model, Fₖ: Forward response at mₖ, Jₖ: Jacobian at mₖ.

MT.mcmc_cache — Type

mutable struct struct mcmc_cache{T1 <: AbstractGeophyModelDistribution, T2 <: AbstractGeophyResponseDistribution}
    apriori::T1
    likelihood::T2
    n_samples::Int
    sampler
end

placeholder to store

apriori in the form of any subtype of AbstractModelDistribution

likelihood in the form of any subtype of AbstractResponseDistribution

number of samples to obtain in n_samples

Turing.jl sampler to be used in sampler

MT.mixing_models — Type

mixing_models

constructs a mixing_models type which can then be used to do rock physics modeling. Should be called using construct_mixing_models

MT.nl_cache — Type

nl_cache: specifies the inverse algorithm while having a cache.

MT.occam_cache — Type

occam_cache: specifies the inverse algorithm while having a cache.

MT.opt_cache — Type

nl_cache: specifies the inverse algorithm while having a cache.

MT.return_code — Type

struct return_code: contains the information if the inversion was successful

MT.rto_cache — Type

rto_cache(m₀, μgrid, alg, max_iters, n_samples, χ2, response_fields, L, verbose)

returns rto_cache that specifies the algorithm to be used for stochastic inversion using RTO-TKO

Arguments

m₀: model, usually a starting value to start inversion but mostly to allocate memory space

μgrid: prior space of μ

alg: type of algorithm for optimization step, choices include (occam_cache)[@ref], (nl_cache)[@ref], and (opt_cache)[@ref]

max_iters: maximum iterations for the optimization scheme

n_samples: number of samples to be obtained. Note: If there are any spurious samples (say NaN values), they will be automatically deleted, reducing the sample size

χ2: target misfit

response_fields: choose data of response to perform inversion on, eg., ρₐ for MT, by default chooses all the data (ρₐ and ϕ)

L: Regularization matrix, defaults to discretized derivative matrix given by ∂(@ref)

verbose: to print results or not

MT.single_phase — Type

single_phase

Single phase only conductivity. Assumes the rock matrix is composed of a single phase only.

MT.transform_utils — Method

transform_utils: Contains the parameters and functions for transformation from optimization to model domains.